Refactor for project specific warnings
